% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_CellPhoneDB_output.R
\name{convert_CellPhoneDB_output}
\alias{convert_CellPhoneDB_output}
\title{Converts CellPhoneDB output matrix into an array of weighted adjacency matrices}
\usage{
convert_CellPhoneDB_output(CellPhoneDB_output, complex_input, gene_input)
}
\arguments{
\item{CellPhoneDB_output}{significant_means.txt file in the output of CellPhoneDB.}

\item{complex_input}{complex_input.csv file. Download from CellPhoneDB.}

\item{gene_input}{gene_input.csv file. Download from CellPhoneDB.}
}
\value{
list of:

              weight_array: 3D numeric array                       array of weighted adjacency matrices with dimensions [number of nodes, number of nodes, number of ligand-receptor pairs].
Note that the function filters out empty weight arrays and the corresponding ligand-receptor pairs
                      First dimension: sending nodes (note: this dimension has all possible nodes, even if some of them are silent for a particular ligand-receptor pair)
                      Second dimension: receiving nodes (note: this dimension has all possible nodes, even if some of them are silent for a particular ligand-receptor pair)
                      Third dimension: ligand-receptor pairs
                      Note that the weight_array should contain dimnames: dimnames = list(nodes, nodes, ligand-receptor pairs)

              ligand_receptor_pair_df: string dataframe          data frame with columns "pair", "ligand", "ligand_complex_composition", "receptor", "receptor_complex_composition"
                       "pair" contains values in a form "ligand:receptor", i.e. ligand being at the first place, receptor being at the second place, e.g. "TNFSF13:TNFRSF17"
                       "ligand" contains ligand names, e.g. "TNFSF13"
                       "ligand_complex_composition" if ligand is a complex (e.g. "aXb2_complex"), contains genes in the ligand complex separated with a comma, e.g. "ITGAX,ITGB2", else contains ""
                       "receptor" contains receptor names, e.g. "TNFRSF17"
                       "receptor_complex_composition" if receptor is a complex (e.g. "NKG2D_II_receptor"), contains genes in the receptor complex separated with a comma, e.g. "KLRK1,HCST", else contains ""
              nodes: character string vector                            a vector with all cell types in the data
}
\description{
\code{convert_CellPhoneDB_output} converts CellPhoneDB output matrix (significant_means.txt) into an array of weighted adjacency matrices.
The CellPhoneDB weight matrix (significant_means.txt) contains ligand-receptor pairs in the rows and all pairs of cell types in the columns.
COMUNET transforms it into a stack of weight matrices (i.e. weight array), one weight matrix per ligand-receptor pair.
If CellPhoneDB output contains \code{m} non-empty ligand-receptor pairs, then the number of matrices in the stack will be equal to \code{m}.
Each such weight matrix has all cell types in the rows and all cell types in the columns.
By this, if there are \code{n} cell types in the data, each weight matrix is an \code{n x n} matrix.
The rows of each matrix represent sending cell types, i.e. cell types that express a ligand of a ligand-receptor pair,
or just partner A of an undirected interacting pair (e.g. pair of adhesion molecules).
The columns of each matrix represent receiving cell types, i.e. cell types that express a receptor of a ligand-receptor pair,
or partner B of an undirected interacting pair (e.g. pair of adhesion molecules).
The NA values of the CellPhoneDB output are substituted with 0.
}
\author{
Maria Solovey \email{maria.solovey@helmholtz-muenchen.de}
}
