% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrp_clustering.R
\name{lrp_clustering}
\alias{lrp_clustering}
\title{Clusters interacting protein pairs}
\usage{
lrp_clustering(
  weight_array,
  ligand_receptor_pair_df,
  nodes,
  dissimilarity = d_normWeightDiff,
  deep_split = 0,
  min_cluster_size = 6
)
}
\arguments{
\item{weight_array}{Numeric array: array of weighted adjacency matrices with dimensions [number of nodes, number of nodes, number of ligand-receptor pairs].}

\item{ligand_receptor_pair_df}{Character string data frame: data frame with columns:

 \itemize{
 \item "pair" contains values in a form "ligand:receptor", i.e. ligand being at the first place, receptor being at the second place, e.g. "TNFSF13:TNFRSF17".
 \item "ligand" contains ligand names, e.g. "TNFSF13".
 \item "ligand_complex_composition" if ligand is a complex (e.g. "aXb2_complex"),
 contains genes in the ligand complex separated with a comma, e.g. "ITGAX,ITGB2", else contains empty string "".
 \item "receptor" contains receptor names, e.g. "TNFRSF17".
 \item "receptor_complex_composition" if receptor is a complex (e.g. "NKG2D_II_receptor"),
 contains genes in the receptor complex separated with a comma, e.g. "KLRK1,HCST", else contains empty string "".
 }}

\item{nodes}{Character string vector: a vector with all cell types in the data.}

\item{dissimilarity}{Function: dissimilarity function. Default value is d_normWeightDiff.}

\item{deep_split}{Numeric: defines the \code{deepSplit} parameter of the \code{\link[cutreeHybrid:dynamicTreeCut]{cutreeHybrid}} function. Default value: 0.}

\item{min_cluster_size}{Numeric: defines the \code{minClusterSize} parameter of the \code{\link[cutreeHybrid:dynamicTreeCut]{cutreeHybrid}} function, i.e. the minimum number of ligand-receptor pairs per cluster. Default value: 6.}
}
\value{
A list of:

\itemize{
  \item{dissim_matrix}{

  Numeric matrix: pairwise dissimilarity between all ligand-receptor pairs.
  }
  \item{clusters}{

  Numeric vector: cluster assignment for each ligand-receptor pair.
  }
  \item{weight_array_by_cluster}{

  Numeric array:   array of weight adjacency matrices (one weight adjacency matrix per cluster).

  The array of weight adjacency has the following dimensions: number of nodes, number of nodes, number of ligand-receptor clusters.

  Each edge weight in a cluster is calculated as the mean of the weights of this edge in all ligand-receptor pairs that belong to this cluster.

  Each delta degree of a node in a cluster is calculated as the mean of the delta degrees of this node in all ligand-receptor pairs that belong to this cluster.
  }
  }
}
\description{
Clusters interacting protein pairs by performing hierarchical clustering of the dissimilarity matrix of layers
(by default, the \code{\link[hclust:stats]{hclust}} R function with the \code{average} agglomeration method is used) and the results can be visualized as a heatmap or a UMAP plot.

The number of clusters is estimated using the  \code{\link[cutreeHybrid:dynamicTreeCut]{cutreeHybrid}} R function
with \code{deepSplit} equal to 0 and default \code{minClusterSize} equal to 6 by default (both can be adjusted).

For each cluster, a graph represents the average communication pattern in the cluster and is built by averaging the edge weights and the delta degrees of the nodes of all the graphs in the cluster.

Please nota that for simplicity, we address all interacting partners (including non-directional partners such as adhesion molecules) as "ligand-receptor pairs".
}
\examples{
# load embryo_interactions data
data("embryo_interactions")

# calculate clusters if ligand-receptor pairs
lrp_clusters <- lrp_clustering(weight_array = embryo_interactions$weight_array,
          ligand_receptor_pair_df = embryo_interactions$ligand_receptor_pair_df,
          nodes = embryo_interactions$nodes
          )

# print(str(lrp_clusters))
}
\author{
Maria Solovey \email{maria.solovey@helmholtz-muenchen.de}
}
