% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_search.R
\name{pattern_search}
\alias{pattern_search}
\title{Performs pattern search}
\usage{
pattern_search(
  pattern_adj_matrix,
  weight_array,
  ligand_receptor_pair_df,
  nodes,
  dissimilarity = d_normWeightDiff
)
}
\arguments{
\item{pattern_adj_matrix}{Numeric matrix: an adjacency matrix for the pattern of interest.}

\item{weight_array}{Numeric array: array of weighted adjacency matrices with dimensions [number of nodes, number of nodes, number of ligand-receptor pairs].}

\item{ligand_receptor_pair_df}{Character dataframe: data frame with columns "pair", "ligand", "ligand_complex_composition", "receptor", "receptor_complex_composition".

Column "pair" contains values in a form "ligand:receptor", i.e. ligand being at the first place, receptor being at the second place, e.g. "TNFSF13:TNFRSF17".

Column "ligand" contains ligand names, e.g. "TNFSF13".

Column "ligand_complex_composition" if ligand is a complex (e.g. "aXb2_complex"),
contains genes in the ligand complex separated with a comma, e.g. "ITGAX,ITGB2", else contains empty string "".

Column "receptor" contains receptor names, e.g. "TNFRSF17".

Column "receptor_complex_composition" if receptor is a complex (e.g. "NKG2D_II_receptor"),
contains genes in the receptor complex separated with a comma, e.g. "KLRK1,HCST", else contains empty string "".}

\item{nodes}{Character string vector: a vector with all cell types in the data.}

\item{dissimilarity}{Function: dissimilarity measure. Default value: d_normWeightDiff.}
}
\value{
dataframe:                              The data frame is sorted by increasing dissimilarity (i.e. similar patterns a the top
              "pair" (character string vector): ligand-receptor pair names
              "dissimilarity" (numeric vector): dissimilarity of ligand-receptor pairs to the pattern of interest. Note here that for the calculation of the dissimilarity of a weight matrix to the pattern of interest, we don't take into consideration the actual weight of the edges in this weight matrix, but just check wether the edge is present (i.e. has a non-zero value) or absent (i.e. has a zero value).
}
\description{
\code{pattern_search} performs a search search for ligand-receptor pairs with a pattern of communication similar to the pattern of interest.
A pattern of interest is defined with a binary adjacency matrix with values 0 or 1, the value 1 indicating the presence of an edge and 0 the absence.
The adjacency matrices of the layers are also binarized based on the presence/absence of the edges and then the dissimilarities with the user-specified adjacency matrix are calculated.
The output is a list of pairs of interacting partners sorted by increasing dissimilarity with the user-specified pattern.
}
\author{
Maria Solovey \email{maria.solovey@helmholtz-muenchen.de}
}
