% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_heatmap.R
\name{plot_cluster_heatmap}
\alias{plot_cluster_heatmap}
\title{Plots heatmap for clustered ligand-receptor pairs}
\usage{
plot_cluster_heatmap(
  dissim_matrix,
  lrp_clusters,
  cluster_colors = "",
  unclustered_LRP_color = "black",
  show_column_names = F,
  show_row_names = F,
  legend_direction = "horizontal",
  legend_width = unit(5, "cm"),
  title_position = "lefttop",
  column_title = "Clustering of ligand-receptor pairs",
  column_title_side = "top",
  heatmap_legend_side = "bottom",
  ...
)
}
\arguments{
\item{dissim_matrix}{numeric matrix: pairwise dissimilarity matrix. Note that rownames and colnames should be defined.}

\item{lrp_clusters}{numeric vector: cluster assignment for each ligand-receptor pair. 
Note that the the vector should be named and ordered in the same way as the rownames and colnames of the dissim_matrix.}

\item{cluster_colors}{character string vector: vector with colours for each cluster. Default value: "".}

\item{unclustered_LRP_color}{character string: colour for ligand-receptor pairs that remained unclustered (cluster assignment 0). Default value: "black".}

\item{show_row_names}{logical: argument of the Heatmap function. Default value: F.}

\item{legend_direction}{character string: argument of the Heatmap function. Default value: "horizontal".}

\item{legend_width}{object of class "unit": argument of the Heatmap function. Default is: unit(5, "cm").}

\item{title_position}{character string: argument of the Heatmap function. Default value: "lefttop".}

\item{column_title}{character string: title of the heatmap plot. Argument of the draw function. Default value: "Clustering of ligand-receptor pairs".}

\item{column_title_side}{character string:  position of the title of the heatmap plot. Argument of the draw function. Default value: "top".}

\item{heatmap_legend_side}{character string: position of the legend of the heatmap plot. Argument of the draw function. Default value: "bottom".}

\item{...}{any arguments "Heatmap" function takes.}

\item{show_colomn_names}{logical: argument of the Heatmap function. Default value: F.}
}
\value{
heatmap with ligand-receptor pairs ordered by their cluster assignment
}
\description{
\code{plot_cluster_heatmap} plots a heatmap for clustered ligand-receptor pairs.
Rows and columns are sorted by cluster number, the color of the heatmap represents dissimilarity between two pairs.
}
\author{
Maria Solovey \email{maria.solovey@helmholtz-muenchen.de}
}
