% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_UMAP.R
\name{plot_cluster_UMAP}
\alias{plot_cluster_UMAP}
\title{Plots UMAP of ligand-receptor pairs}
\usage{
plot_cluster_UMAP(
  ligand_receptor_pair_df,
  dissim_matrix,
  lrp_clusters,
  seed = 100,
  cluster_colors = "",
  unclustered_LRP_color = "black",
  cluster_shape = "",
  point_size = 2,
  title = "umap of ligand-receptor pairs",
  legend_position = "bottom",
  legend_direction = "horizontal",
  legend_title_position = "top",
  legend_label_position = "bottom",
  legend_byrow = TRUE
)
}
\arguments{
\item{ligand_receptor_pair_df}{Character dataframe: data frame with columns "pair", "ligand", "ligand_complex_composition", "receptor", "receptor_complex_composition".

Column "pair" contains values in a form "ligand:receptor", i.e. ligand being at the first place, receptor being at the second place, e.g. "TNFSF13:TNFRSF17".

Column "ligand" contains ligand names, e.g. "TNFSF13".

Column ligand_complex_composition" if ligand is a complex (e.g. "aXb2_complex"),
contains genes in the ligand complex separated with a comma, e.g. "ITGAX,ITGB2", else contains empty string "".

Column "receptor" contains receptor names, e.g. "TNFRSF17".

Column "receptor_complex_composition" if receptor is a complex (e.g. "NKG2D_II_receptor"),
contains genes in the receptor complex separated with a comma, e.g. "KLRK1,HCST", else contains empty string "".}

\item{dissim_matrix}{Numeric matrix: pairwise dissimilarity matrix. Note that rownames and colnames should be defined.}

\item{lrp_clusters}{Numeric vector: cluster assignment for each ligand-recetor pair.
Note that the the vector should be named and ordered in the same way as the rownames and colnames of the dissim_matrix.}

\item{seed}{Random seed for UMAP. Default value: 100.}

\item{cluster_colors}{Character string vector: vector with colours for each cluster. Default value: "".}

\item{unclustered_LRP_color}{Character string: colour for ligand-receptor pairs that remained unclustered (cluster assignment 0). Default value: "black".}

\item{cluster_shape}{Numeric vector: vector with shapes for each cluster. Default value: "".}

\item{point_size}{Numeric: size of points. Default value: 2.}

\item{title}{Character string: title of the heatmap plot. Argument of the draw function. Default value: "umap of ligand-receptor pairs".}

\item{legend_position}{Character string: argument of the theme function. Default value: "bottom" guide_legend.}

\item{legend_direction}{Character string: argument of the guide_legend function. Default value: "horizontal".}

\item{legend_title_position}{Character string: argument of the guide_legend function. Default value: "top".}

\item{legend_label_position}{Character string: argument of the guide_legend function. Default value: "bottom".}

\item{legend_byrow}{logical: Argument of the guide_legend function. Default value: TRUE.}
}
\value{
UMAP of ligand-receptor pairs coloured and shaped by cluster
}
\description{
\code{plot_cluster_UMAP} plots UMAP of ligand-receptor pairs colored and shaped by cluster.
}
\author{
Maria Solovey \email{maria.solovey@helmholtz-muenchen.de}
}
