% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lig_rec.R
\name{plot_lig_rec}
\alias{plot_lig_rec}
\title{Plots interacting partners in a particular cluster}
\usage{
plot_lig_rec(
  cluster_of_interest,
  lrp_clusters,
  ligand_receptor_pair_df,
  lig_rec_color = c("red", "blue"),
  edge_arrow_size = 0.25,
  node_label_cex = 1,
  layout = layout_nicely,
  legend_position = c(-1, -1.1),
  legend_pt_cex = 2,
  legend_cex = 0.8,
  ...
)
}
\arguments{
\item{cluster_of_interest}{Integer: number of the cluster, for which the ligands and receptors should be plotted.}

\item{lrp_clusters}{Integer vector: cluster assignment for each ligand-receptor pair.}

\item{ligand_receptor_pair_df}{Character string data frame: data frame with columns:

 \itemize{
 \item "pair" contains values in a form "ligand:receptor", i.e. ligand being at the first place, receptor being at the second place, e.g. "TNFSF13:TNFRSF17".
 \item "ligand" contains ligand names, e.g. "TNFSF13".
 \item "ligand_complex_composition" if ligand is a complex (e.g. "aXb2_complex"),
 contains genes in the ligand complex separated with a comma, e.g. "ITGAX,ITGB2", else contains empty string "".
 \item "receptor" contains receptor names, e.g. "TNFRSF17".
 \item "receptor_complex_composition" if receptor is a complex (e.g. "NKG2D_II_receptor"),
 contains genes in the receptor complex separated with a comma, e.g. "KLRK1,HCST", else contains empty string "".
 }}

\item{lig_rec_color}{Character string vector: colours for ligand and receptors: first colour for the ligand, second colour for the receptor. Default value: c("red", "blue").}

\item{edge_arrow_size}{Numeric: size of arrow edge. Default value: 0.25. See also \code{edge.arrow.size} argument of the \code{\link[plot.igraph:igraph]{plot.igraph}} function.}

\item{node_label_cex}{Numeric: size of node labels. Default value: 1.}

\item{layout}{Function: see layout from igraph package v0.2.1. Default value: layout_nicely.}

\item{legend_position}{Numeric vector: x, y coordinates of the legend position. Default value: c(-1, -1.1).}

\item{legend_pt_cex}{Numeric: expansion factor(s) for the points in the legend. Default value: 2. See also \code{pt.cex} argument of the \code{\link[legend:graphics]{legend}} function.}

\item{legend_cex}{Numeric: character expansion factor in the legend. Default value: 0.8. See also \code{cex} argument of the \code{\link[legend:graphics]{legend}} function.}

\item{...}{Any other \code{\link[plot.igraph:igraph]{plot.igraph}} parameters.}
}
\value{
Graph plot.
}
\description{
Plots interacting partners in a particular cluster.

If the pair of interacting partners is a ligand-receptor pair, then the ligand is colored red,
the receptor is colored blue, and the arrow goes from the ligand to the receptor.
In case no directionality is specified for the pair of interacting proteins A and B (i.e., as for adhesion molecules),
the arrows start at the node expressing partner A and point to the node expressing partner B.

Please nota that for simplicity, we address all interacting partners (including non-directional partners such as adhesion molecules) as "ligand-receptor pairs".
}
\examples{

# load embryo_interactions
data("embryo_interactions")

# load lrp_clusters
data("embryo_lrp_clusters")

plot_lig_rec(cluster_of_interest = 1,
             lrp_clusters = embryo_lrp_clusters$clusters,
             ligand_receptor_pair_df = embryo_interactions$ligand_receptor_pair_df)

}
\author{
Maria Solovey \email{maria.solovey@helmholtz-muenchen.de}
}
