% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparative_analysis.R
\name{comparative_analysis}
\alias{comparative_analysis}
\title{Performs comparative analysis between two conditions}
\usage{
comparative_analysis(
  cond1_weight_array,
  cond2_weight_array,
  cond1_ligand_receptor_pair_df,
  cond2_ligand_receptor_pair_df,
  cond1_nodes,
  cond2_nodes = cond1_nodes,
  cond1_name,
  cond2_name,
  dissimilarity = d_normWeightDiff
)
}
\arguments{
\item{cond1_weight_array}{Numeric array: array of weighted adjacency matrices for condition 1.

Note that the function filters out empty weight arrays and the corresponding ligand-receptor pairs.}

\item{cond2_weight_array}{Numeric array: array of weighted adjacency matrices for condition 2.

Note that the function filters out empty weight arrays and the corresponding ligand-receptor pairs.}

\item{cond1_ligand_receptor_pair_df}{Character data frame: data frame with columns "pair", "ligand", "ligand_complex_composition", "receptor", "receptor_complex_composition" for condition 1.}

\item{cond2_ligand_receptor_pair_df}{Character data frame: data frame with columns "pair", "ligand", "ligand_complex_composition", "receptor", "receptor_complex_composition" for condition 2.}

\item{cond1_nodes}{Character string vector: a vector with all cell types in the data.

Note that comparative_analysis expects nodes (i.e. cell types) to be the same in both conditions.}

\item{cond2_nodes}{Character string vector: a vector with all cell types in the data. Default value: cond1_nodes.

Note that comparative_analysis expects nodes (i.e. cell types) to be the same in both conditions.}

\item{cond1_name}{Character string: sample name for condition 1.}

\item{cond2_name}{Character string: sample name for condition 2.}

\item{dissimilarity}{Function: dissimilarity function. Default value is d_normWeightDiff.}
}
\value{
A list of:

\itemize{
  \item{sorted_LRP_df}{

  Data frame with columns:
  \itemize{
    \item "pair" (character string vector): names of ligand-receptor pairs in the same form as they are in ligand_receptor_pair_df$pair.
    \item "presence" (character string vector): whether the ligand-receptor pair is present in both conditions ("shared") or only in one of them.
    \item "dissimilarity" (numeric vector): dissimilarity value between the topology of ligand-receptor pair graph in two conditions.
    The smaller the dissimilarity value, the more similar is the graph topology between the two conditions.
    If a ligand-receptor pair is present only in one of the conditions, the dissimilarity is equal to 1.

  }
  }
  \item{dissim_cond1_cond2}{

  Numeric matrix:
   pairwise dissimilarity between all ligand-receptor pairs in the two conditions (condition 1 in rows, condition 2 in columns).
  }
}
}
\description{
Performs comparative analysis between two conditions (e.g. two samples).
Given the results of a cell-cell communication analysis of two datasets including the same cell types,
\code{comparative_analysis} estimates the differences in the cell-cell communication patterns between corresponding interacting partners in the two datasets.

Please note that for simplicity, we address all interacting partners (including non-directional partners such as adhesion molecules) as "ligand-receptor pairs".
}
\examples{
# load AML328_d0_interactions
data("AML328_d0_interactions")

# load AML328_d29_interactions
data("AML328_d29_interactions")

my_comparative_analysis <- comparative_analysis(cond1_weight_array = AML328_d0_interactions$weight_array
                                             ,cond2_weight_array = AML328_d29_interactions$weight_array
                                             ,cond1_ligand_receptor_pair_df = AML328_d0_interactions$ligand_receptor_pair_df
                                             ,cond2_ligand_receptor_pair_df = AML328_d29_interactions$ligand_receptor_pair_df
                                             ,cond1_nodes = AML328_d0_interactions$nodes
                                             ,cond2_nodes = AML328_d29_interactions$nodes
                                             ,cond1_name = "AML328_d0"
                                             ,cond2_name = "AML328_d29")
print(str(my_comparative_analysis))

}
\author{
Maria Solovey \email{maria.solovey@helmholtz-muenchen.de}
}
