% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrp_clustering.R
\name{lrp_clustering}
\alias{lrp_clustering}
\title{Clusters interacting protein pairs}
\usage{
lrp_clustering(
  weight_array,
  ligand_receptor_pair_df,
  nodes,
  dissimilarity = d_normWeightDiff,
  deep_split = 0,
  min_cluster_size = 6
)
}
\arguments{
\item{weight_array}{numeric array: array of weighted adjacency matrices with dimensions [number of nodes, number of nodes, number of ligand-receptor pairs].}

\item{ligand_receptor_pair_df}{character dataframe: data frame with columns "pair", "ligand", "ligand_complex_composition", "receptor", "receptor_complex_composition".

Column "pair" contains values in a form "ligand:receptor", i.e. ligand being at the first place, receptor being at the second place, e.g. "TNFSF13:TNFRSF17".

Column "ligand" contains ligand names, e.g. "TNFSF13".

Column ligand_complex_composition" if ligand is a complex (e.g. "aXb2_complex"),
contains genes in the ligand complex separated with a comma, e.g. "ITGAX,ITGB2", else contains empty string "".

Column "receptor" contains receptor names, e.g. "TNFRSF17".

Column "receptor_complex_composition" if receptor is a complex (e.g. "NKG2D_II_receptor"),
contains genes in the receptor complex separated with a comma, e.g. "KLRK1,HCST", else contains empty string "".}

\item{nodes}{character string vector: a vector with all cell types in the data.}

\item{dissimilarity}{function: dissimilarity function. Default value is d_normWeightDiff.}

\item{deep_split}{numeric: defines the deepSplit parameter of the \code{cutreeHybrid} function. Default value: 0.}

\item{min_cluster_size}{numeric: minimum number of ligand-receptor pairs per cluster. Default value: 6.}
}
\value{
a list of:

              dissim_matrix: numeric matrix           pairwise dissimilarity between all ligand-receptor pairs

              clusters: numeric vector                cluster assignment for each ligand-receptor pair

              weight_array_by_cluster: numeric array   array of weighted adjacency matrices with dimensions [number of nodes, number of nodes, number of ligand-receptor clusters].
Each edge weight in a cluster is calculated as the mean of the weights of this edge in all ligand-receptor pairs that belong to this cluster.
}
\description{
\code{lrp_clustering} clusters interacting protein pairs by performing hierarchical clustering of the dissimilarity matrix of layers
(by default, the \code{hclust} R function with the \code{average} agglomeration method is used) and the results can be visualized as a heatmap or a UMAP plot (McInnes et al., 2018).
The number of clusters is estimated using the  \code{cutreeHybrid} R function (package dynamicTreeCut, version 1.63-1(Langfelder et al., 2008))
with \code{deepSplit} equal to 0 and default \code{minClusterSize} equal to 6 by default (both can be adjusted).
For each cluster, a graph that represents the “average” pattern is built by averaging the adjacency matrices and the (see section 2.1) of the nodes of all the graphs in the cluster.
}
\author{
Maria Solovey \email{maria.solovey@helmholtz-muenchen.de}
}
