% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pattern_matrix.R
\name{make_pattern_matrix}
\alias{make_pattern_matrix}
\title{Makes a weight matrix for pattern of interest}
\usage{
make_pattern_matrix(communicating_nodes, nodes)
}
\arguments{
\item{communicating_nodes}{Character string vector: in form of c("cellType1_to_cellType2", "cellType1_to_cellType3", ...), e.g. c("HSC_to_Tcells", "HSC_to_Bcells", ...).}

\item{nodes}{Character string vector: of length n with all cell types in the data.}
}
\value{
\item{weight_matrix}{

  Numeric \eqn{n} x \eqn{n} matrix with \eqn{n} being the number of nodes and the values of the matrix being 0 or 1.
  Rows are regarded as sending cell type, columns are regarded as receiving cell types. A value of 1 defines communication, a value of 0 defines no communication.
}
}
\description{
Makes a weight matrix for pattern of interest.

The rows and columns of the matrix contain all nodes.
The rows are considered as sending nodes and the columns are considered as receiving nodes.
The value of 1 represents an edge (i.e. cell types are communicating) and 0 represents no edge (i.e. cell types are not communicating).

Please nota that for simplicity, we address all interacting partners (including non-directional partners such as adhesion molecules) as "ligand-receptor pairs".
}
\examples{

# load embryo_interactions
data("embryo_interactions")

communicating_nodes <- c("exVE_to_EPI" ,"exVE_to_Mes","exVE_to_TE" ,"exVE_to_emVE" ,"exVE_to_exVE")

test_pattern <- make_pattern_matrix(communicating_nodes = communicating_nodes,
                   nodes = embryo_interactions$nodes)
print(test_pattern)


}
\author{
Maria Solovey \email{maria.solovey@helmholtz-muenchen.de}
}
