% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_search.R
\name{pattern_search}
\alias{pattern_search}
\title{Performs pattern search}
\usage{
pattern_search(
  pattern_adj_matrix,
  weight_array,
  ligand_receptor_pair_df,
  nodes,
  dissimilarity = d_normWeightDiff
)
}
\arguments{
\item{pattern_adj_matrix}{Numeric matrix: an adjacency matrix for the pattern of interest.}

\item{weight_array}{Numeric array: array of weighted adjacency matrices with dimensions [number of nodes, number of nodes, number of ligand-receptor pairs].}

\item{ligand_receptor_pair_df}{Character string data frame: data frame with columns:

 \itemize{
 \item "pair" contains values in a form "ligand:receptor", i.e. ligand being at the first place, receptor being at the second place, e.g. "TNFSF13:TNFRSF17".
 \item "ligand" contains ligand names, e.g. "TNFSF13".
 \item "ligand_complex_composition" if ligand is a complex (e.g. "aXb2_complex"),
 contains genes in the ligand complex separated with a comma, e.g. "ITGAX,ITGB2", else contains empty string "".
 \item "receptor" contains receptor names, e.g. "TNFRSF17".
 \item "receptor_complex_composition" if receptor is a complex (e.g. "NKG2D_II_receptor"),
 contains genes in the receptor complex separated with a comma, e.g. "KLRK1,HCST", else contains empty string "".
 }}

\item{nodes}{Character string vector: a vector with all cell types in the data.}

\item{dissimilarity}{Function: dissimilarity measure. Default value: d_normWeightDiff.}
}
\value{
A data frame with columns:
  \itemize{
    \item "pair" (character string vector): ligand-receptor pair names.
    \item "dissimilarity" (numeric vector): dissimilarity of ligand-receptor pairs to the pattern of interest.
    Note here that for the calculation of the dissimilarity of a weight matrix to the pattern of interest,
    we don't take into consideration the actual weight of the edges in this weight matrix,
    but just check wether the edge is present (i.e. has a non-zero value) or absent (i.e. has a zero value).

  }
   The data frame is sorted by increasing dissimilarity (i.e. similar patterns come at the top).
}
\description{
Performs a search search for interacting partners with a pattern of communication similar to the pattern of interest.

A pattern of interest is defined with a binary adjacency matrix with values 0 or 1, the value 1 indicating the presence of an edge and 0 the absence.
The adjacency matrices of the layers are also binarized based on the presence/absence of the edges and then the dissimilarities with the user-specified adjacency matrix are calculated.

Please note that for simplicity, we address all interacting partners (including non-directional partners such as adhesion molecules) as "ligand-receptor pairs".
}
\examples{
# load embryo_interactions
data("embryo_interactions")

# make a pattern of interest
communicating_nodes <- c("exVE_to_EPI" ,"exVE_to_Mes","exVE_to_TE" ,"exVE_to_emVE" ,"exVE_to_exVE")

test_pattern <- make_pattern_matrix(communicating_nodes = communicating_nodes,
                   nodes = embryo_interactions$nodes)
print(test_pattern)

# alternatively create a pattern matrix manually:
\dontrun{
test_pattern <- matrix(c(rep(0
                            ,dim(embryo_interactions$weight_array)[[1]]-1)
                            ,1)
                       ,nrow = dim(embryo_interactions$weight_array)[[1]]
                       ,ncol = dim(embryo_interactions$weight_array)[[2]]
                       ,byrow = F
                       )
dimnames(test_pattern) <- dimnames(embryo_interactions$weight_array)[c(1,2)]
}

# search for the pattern
patterns <- pattern_search(pattern_adj_matrix = test_pattern,
                          weight_array = embryo_interactions$weight_array,
                          ligand_receptor_pair_df = embryo_interactions$ligand_receptor_pair_df,
                          nodes = embryo_interactions$nodes)
print(str(patterns))

}
\author{
Maria Solovey \email{maria.solovey@helmholtz-muenchen.de}
}
