% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_heatmap.R
\name{plot_cluster_heatmap}
\alias{plot_cluster_heatmap}
\title{Plots heatmap for clustered ligand-receptor pairs}
\usage{
plot_cluster_heatmap(
  dissim_matrix,
  lrp_clusters,
  cluster_colors = "",
  unclustered_LRP_color = "black",
  show_column_names = FALSE,
  show_row_names = FALSE,
  legend_direction = "horizontal",
  legend_width = unit(5, "cm"),
  title_position = "lefttop",
  column_title = "Clustering of ligand-receptor pairs",
  column_title_side = "top",
  heatmap_legend_side = "bottom",
  ...
)
}
\arguments{
\item{dissim_matrix}{Numeric matrix: pairwise dissimilarity matrix.

Note that row names and column names should be defined.}

\item{lrp_clusters}{Numeric vector: cluster assignment for each ligand-receptor pair.

Note that the the vector should be named and ordered in the same way as the row names and column names of the dissim_matrix.}

\item{cluster_colors}{Character string vector: vector with colours for each cluster. Default value: "".}

\item{unclustered_LRP_color}{Character string: colour for ligand-receptor pairs that remained unclustered (cluster assignment 0). Default value: "black".}

\item{show_row_names}{Logical: argument of the \code{\link[Heatmap:ComplexHeatmap]{Heatmap}} function. Default value: FALSE.}

\item{legend_direction}{Character string: argument of the \code{\link[Heatmap:ComplexHeatmap]{Heatmap}} function. Default value: "horizontal".}

\item{legend_width}{Object of class "unit": argument of the \code{\link[Heatmap:ComplexHeatmap]{Heatmap}} function. Default is: unit(5, "cm").}

\item{title_position}{Character string: argument of the \code{\link[Heatmap:ComplexHeatmap]{Heatmap}} function. Default value: "lefttop".}

\item{column_title}{Character string: title of the heatmap plot. Argument of the \code{\link[draw:ComplexHeatmap]{draw}} function. Default value: "Clustering of ligand-receptor pairs".}

\item{column_title_side}{Character string:  position of the title of the heatmap plot. Argument of the \code{\link[draw:ComplexHeatmap]{draw}} function. Default value: "top".}

\item{heatmap_legend_side}{Character string: position of the legend of the heatmap plot. Argument of the \code{\link[draw:ComplexHeatmap]{draw}} function. Default value: "bottom".}

\item{...}{Any arguments \code{\link[Heatmap:ComplexHeatmap]{Heatmap}} function takes.}

\item{show_colomn_names}{Logical: argument of the \code{\link[Heatmap:ComplexHeatmap]{Heatmap}} function. Default value: FALSE.}
}
\value{
Heatmap with ligand-receptor pairs ordered by their cluster assignment.
}
\description{
Plots a heatmap for clustered ligand-receptor pairs.

Rows and columns are sorted by cluster number, the color of the heatmap represents dissimilarity between two ligand-receptor pairs.
}
\examples{

# load lrp_clusters
data("lrp_clusters")

# plot heatmap
plot_cluster_heatmap(dissim_matrix = lrp_clusters$dissim_matrix, lrp_clusters = lrp_clusters$clusters)

}
\author{
Maria Solovey \email{maria.solovey@helmholtz-muenchen.de}
}
