% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_UMAP.R
\name{plot_cluster_UMAP}
\alias{plot_cluster_UMAP}
\title{Plots UMAP of ligand-receptor pairs}
\usage{
plot_cluster_UMAP(
  ligand_receptor_pair_df,
  dissim_matrix,
  lrp_clusters,
  seed = 100,
  cluster_colors = "",
  unclustered_LRP_color = "black",
  cluster_shape = "",
  point_size = 2,
  title = "umap of ligand-receptor pairs",
  legend_position = "bottom",
  legend_direction = "horizontal",
  legend_title_position = "top",
  legend_label_position = "bottom",
  legend_byrow = TRUE
)
}
\arguments{
\item{ligand_receptor_pair_df}{Character string data frame: data frame with columns:

 \itemize{
 \item "pair" contains values in a form "ligand:receptor", i.e. ligand being at the first place, receptor being at the second place, e.g. "TNFSF13:TNFRSF17".
 \item "ligand" contains ligand names, e.g. "TNFSF13".
 \item "ligand_complex_composition" if ligand is a complex (e.g. "aXb2_complex"),
 contains genes in the ligand complex separated with a comma, e.g. "ITGAX,ITGB2", else contains empty string "".
 \item "receptor" contains receptor names, e.g. "TNFRSF17".
 \item "receptor_complex_composition" if receptor is a complex (e.g. "NKG2D_II_receptor"),
 contains genes in the receptor complex separated with a comma, e.g. "KLRK1,HCST", else contains empty string "".
 }}

\item{dissim_matrix}{Numeric matrix: pairwise dissimilarity matrix.

Note that row names and column names should contain ligand-receptor pairs.}

\item{lrp_clusters}{Numeric vector: cluster assignment for each ligand-recetor pair.

Note that the the vector should be named and ordered in the same way as the row names and column names of the dissim_matrix.}

\item{seed}{Random seed for UMAP. Default value: 100.}

\item{cluster_colors}{Character string vector: vector with colours for each cluster. Default value: "".}

\item{unclustered_LRP_color}{Character string: colour for ligand-receptor pairs that remained unclustered (cluster assignment 0). Default value: "black".}

\item{cluster_shape}{Numeric vector: vector with shapes for each cluster. Default value: "".}

\item{point_size}{Numeric: size of points. Default value: 2.}

\item{title}{Character string: title of the UMAP plot. Argument of the \code{\link[ggtitle:ggplot2]{ggtitle}} function. Default value: "umap of ligand-receptor pairs".}

\item{legend_position}{Character string: argument of the \code{\link[theme:ggplot2]{theme}} function. Default value: "bottom" guide_legend.}

\item{legend_direction}{Character string: argument of the \code{\link[guide_legend:ggplot2]{guide_legend}} function. Default value: "horizontal".}

\item{legend_title_position}{Character string: argument of the \code{\link[guide_legend:ggplot2]{guide_legend}} function. Default value: "top".}

\item{legend_label_position}{Character string: argument of the \code{\link[guide_legend:ggplot2]{guide_legend}} function. Default value: "bottom".}

\item{legend_byrow}{logical: Argument of the \code{\link[guide_legend:ggplot2]{guide_legend}} function. Default value: TRUE.}
}
\value{
UMAP of ligand-receptor pairs coloured and shaped by cluster.
}
\description{
Plots UMAP of ligand-receptor pairs colored and shaped by cluster.
}
\examples{
# load lrp_clusters
data("embryo_lrp_clusters")

# load embryo_interactions
data("embryo_interactions")

# plot heatmap
plot_cluster_UMAP(ligand_receptor_pair_df = embryo_interactions$ligand_receptor_pair_df, dissim_matrix = embryo_lrp_clusters$dissim_matrix, lrp_clusters = embryo_lrp_clusters$clusters)

}
\author{
Maria Solovey \email{maria.solovey@helmholtz-muenchen.de}
}
