% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lig_rec.R
\name{plot_lig_rec}
\alias{plot_lig_rec}
\title{Plots ligands and receptors in a particular cluster}
\usage{
plot_lig_rec(
  cluster_of_interest,
  lrp_clusters,
  ligand_receptor_pair_df,
  lig_rec_color = c("red", "blue"),
  edge_arrow_size = 0.25,
  node_label_cex = 1,
  layout = layout_nicely,
  legend_position = c(-1, -1.1),
  legend_pt_cex = 2,
  legend_cex = 0.8,
  ...
)
}
\arguments{
\item{cluster_of_interest}{numeric: number of the cluster, for which the ligands and receptors should be plotted.}

\item{lrp_clusters}{numeric vector: cluster assignment for each ligand-receptor pair.}

\item{ligand_receptor_pair_df}{character dataframe: data frame with columns "pair", "ligand", "ligand_complex_composition", "receptor", "receptor_complex_composition".

Column "pair" contains values in a form "ligand:receptor", i.e. ligand being at the first place, receptor being at the second place, e.g. "TNFSF13:TNFRSF17".

Column "ligand" contains ligand names, e.g. "TNFSF13".

Column "ligand_complex_composition" if ligand is a complex (e.g. "aXb2_complex"),
contains genes in the ligand complex separated with a comma, e.g. "ITGAX,ITGB2", else contains empty string "".

Column "receptor" contains receptor names, e.g. "TNFRSF17".

Column "receptor_complex_composition" if receptor is a complex (e.g. "NKG2D_II_receptor"),
contains genes in the receptor complex separated with a comma, e.g. "KLRK1,HCST", else contains empty string "".}

\item{lig_rec_color}{character string vector: colours for ligand and receptors: first colour for the ligand, second colour for the receptor. Default value: c("red", "blue").}

\item{edge_arrow_size}{numeric: size of arrow edge. Default value: 0.25. See same argument of plot.igraph().}

\item{node_label_cex}{numeric: size of node labels. Default value: 1.}

\item{layout}{function: see layout from igraph package v0.2.1. Default value: layout_nicely.}

\item{legend_position}{numeric vector: x, y coordinates of the legend position. Default value: c(-1, -1.1).}

\item{legend_pt_cex}{numeric: expansion factor(s) for the points in the legend. See legend() from graphics v3.6.2. Default value: 2.}

\item{legend_cex}{numeric: character expansion factor in the legend. See legend() from graphics v3.6.2. Default value: 0.8.}

\item{...}{other plot.igraph parameters}
}
\value{
graph plot
}
\description{
\code{plot_lig_rec} plots interacting partners in a particular cluster.
If the pair of interacting partners is a ligand-receptor pair, then the ligand is colored red,
the receptor is colored blue, and the arrow goes from the ligand to the receptor.
In case no directionality is specified for the pair of interacting proteins A and B (i.e., as for adhesion molecules),
the arrows start at the node expressing partner A and point to the node expressing partner B.
}
\author{
Maria Solovey \email{maria.solovey@helmholtz-muenchen.de}
}
